feat(native): Phase 4 — pyo3-free fingerprint-core; pgrx computes in pure Rust
Extracts the canonical fingerprint canonicalizer into a new pyo3-free / pgrx-free
crate (goldenmatch-fingerprint-core) shared by every surface, and rewires the
Postgres extension to compute the fingerprint IN-PROCESS instead of through the
embedded-CPython bridge -- the first SQL function that needs no interpreter for
its work (the long-game "decouple SQL extensions from embedded CPython" lever).

- fingerprint-core: FpValue + fingerprint_fields + fingerprint_json + sha2/serde
  deps, with unit tests pinning the canonical vectors. Standalone workspace so
  both the `native` (own workspace) and `postgres` (excluded) crates can path-dep
  it cleanly.
- native: hash.rs now maps PyDict -> core::FpValue (pyo3 wrapper) and the C ABI
  calls core::fingerprint_json; the duplicated logic + sha2/serde_json deps move
  to core. Canonical bytes unchanged (pinned golden vectors still hold).
- postgres: goldenmatch_record_fingerprint now calls
  goldenmatch_fingerprint_core::fingerprint_json directly (no bridge, no GIL),
  with a #[pg_test] asserting the shared pinned vector.
- bridge: drop the now-unused record_fingerprint (added in #520 for pgrx; pgrx no
  longer routes through CPython for it).
- ci: native lane gains the fingerprint-core path filter + a clippy+test step;
  the pgrx pg_test validates the pure-Rust path in the rust_pgrx lanes.

No behavior change: the fingerprint value is identical across Python, the native
C ABI, DuckDB, and Postgres -- pgrx just stops embedding an interpreter to get it.
