+++ packages/rust/extensions/native/Cargo.toml
@@ -25,13 +25,9 @@ pyo3 = { version = ">=0.23.3, <0.25", features = ["extension-module", "abi3-py31
 # BLAKE2b for the char-n-gram feature hashing kernel — must match Python
 # hashlib.blake2b(digest_size=8) byte-for-byte (see featurize.rs).
 blake2 = "0.10"
-# SHA-256 for the canonical record fingerprint kernel — must match Python
-# hashlib.sha256 byte-for-byte (see hash.rs + core/_hashing.py).
-sha2 = "0.10"
-# JSON parsing for the gm_record_fingerprint C ABI (non-Python callers pass a
-# JSON object). arbitrary_precision preserves integer literals exactly so big
-# ints match the pyo3 path's Python-str() ints.
-serde_json = { version = "1", features = ["arbitrary_precision"] }
+# Canonical record fingerprint canonicalizer (sha256 + JSON parsing live here),
+# shared pyo3-free with the pgrx `postgres` crate so both compute the same id.
+goldenmatch-fingerprint-core = { path = "../fingerprint-core" }
 rapidfuzz = "0.5.0"
 rayon = "1.12.0"
 # Arrow C Data Interface for zero-copy kernel input (see score.rs
