+++ packages/rust/extensions/postgres/Cargo.toml
@@ -21,6 +21,9 @@ pg_test = []
 [dependencies]
 pgrx = "0.12.9"
 goldenmatch-bridge = { path = "../bridge" }
+# Pure-Rust canonical fingerprint — computed in-process, NOT through the
+# embedded-CPython bridge (the decoupling lever; see kernels.rs).
+goldenmatch-fingerprint-core = { path = "../fingerprint-core" }
 serde_json = "1"
 
 [[bin]]
